+++ libs/ui-card/src/components/left-photo/left-photo-title/left-photo-title.variants.tsx
@@ -1,0 +1,13 @@
+import { VariantProps, createVariants } from '@r7/ui-base-components'
+
+export const variants = createVariants({
+  variants: {
+    fontSize: {
+      small: 'card-text-xxxs card-leading-xs',
+      medium: 'card-text-xxs card-leading-md',
+      large: 'card-text-xxs card-leading-md md:card-text-sm lg:card-leading-lg',
+    },
+  },
+})
+
+export type TitleVariants = VariantProps<typeof variants>