+++ libs/ui-card/src/components/proportional-photo/proportional-photo.stories.tsx
@@ -13,30 +13,30 @@ export default meta
 
 type Story = StoryObj<typeof ProportionalPhoto>
 
-export const Primary: Story = {
-  render: args => (
-    <ProportionalPhoto
-      renderImage={args.renderImage}
-      renderHat={args.renderHat}
-      title={args.title}
-    />
+export const Responsive: Story = {
+  render: () => (
+    <ProportionalPhoto>
+      <ProportionalPhoto.Figure>
+        <img
+          src="//img.r7.com/images/apagao-sao-paulo-1123-19122023115816951?resize=348x199&amp;crop=1170x669 0 32&amp;dimensions=328x184"
+          alt="Ministério da Justiça abre processo contra Enel por apagões"
+          width="100%"
+          height="auto"
+        />
+      </ProportionalPhoto.Figure>
+
+      <ProportionalPhoto.TextWrapper>
+        <Card.HatWrapper>
+          <Card.HatTitle>são paulo e rio de janeiro</Card.HatTitle>
+        </Card.HatWrapper>
+
+        <ProportionalPhoto.Title>
+          Ministério da Justiça abre processo contra Enel por apagões
+        </ProportionalPhoto.Title>
+      </ProportionalPhoto.TextWrapper>
+    </ProportionalPhoto>
   ),
-  args: {
-    renderHat: (
-      <Card.HatWrapper>
-        <Card.HatTitle>são paulo e rio de janeiro</Card.HatTitle>
-      </Card.HatWrapper>
-    ),
-    title: 'Ministério da Justiça abre processo contra Enel por apagões',
-    renderImage: (
-      <img
-        src="//img.r7.com/images/apagao-sao-paulo-1123-19122023115816951?resize=348x199&amp;crop=1170x669 0 32&amp;dimensions=328x184"
-        alt="Ministério da Justiça abre processo contra Enel por apagões"
-        width="100%"
-        height="auto"
-      />
-    ),
-  },
+  args: {},
   parameters: {
     design: {
       type: 'figma',
@@ -44,4 +44,37 @@ export const Primary: Story = {
       accessToken: 'figd_sHs6Ap894w4C-OAFls7tuq0fMMROyue-8zQJ8hRE',
     },
   },
+}
+
+export const Big: Story = {
+  ...Responsive,
+  decorators: [
+    Story => (
+      <div className="card-w-full card-max-w-[350px]">
+        <Story />
+      </div>
+    ),
+  ],
+}
+
+export const Medium: Story = {
+  ...Responsive,
+  decorators: [
+    Story => (
+      <div className="card-w-full card-max-w-[255px]">
+        <Story />
+      </div>
+    ),
+  ],
+}
+
+export const Small: Story = {
+  ...Responsive,
+  decorators: [
+    Story => (
+      <div className="card-w-full card-max-w-[207px]">
+        <Story />
+      </div>
+    ),
+  ],
 }